IMP - Refactor badge component files
