+++ src/components/elements/gameData/badges/conditions.tsx
@@ -4,14 +4,13 @@ import Badge from 'react-bootstrap/Badge';
 import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
 
-import {useI18n} from '../../../i18n/hook';
-import {useAtkSkillResources} from '../../hooks/atkSkillResources';
+import {useI18n} from '../../../../i18n/hook';
+import {useAtkSkillResources} from '../../../hooks/atkSkillResources';
 
 
 type ConditionBadgeProps = {
   conditionCodes: Array<number>,
 }
-
 export const getConditionBadges = ({conditionCodes}: ConditionBadgeProps) => {
   const {conditionEnumMap} = useAtkSkillResources({toFetch: 'conditionEnumsOnly'});
   const {lang} = useI18n();
