+++ src/components/pages/gameData/unitInfo/output/elements/normalAttack/branchedTab.tsx
@@ -4,7 +4,7 @@ import {NormalAttackBranchedChain} from '../../../../../../../api-def/resources'
 import {ConditionCodes} from '../../../../../../../const/gameData';
 import {useI18n} from '../../../../../../../i18n/hook';
 import {sum} from '../../../../../../../utils/calc';
-import {ConditionBadges} from '../../../../../../elements/gameData/conditionBadges';
+import {ConditionBadges} from '../../../../../../elements/gameData/badges/conditions';
 import {Markdown} from '../../../../../../elements/markdown/main';
 import styles from '../main.module.css';
 import {SectionSubTitle} from '../title';
@@ -40,77 +40,73 @@ export const NormalAttackBranchedTab = ({branchedChain}: Props) => {
             </tr>
           </thead>
           <tbody>
-            {
-              branchedChain.combos.map((combo, idx) => {
-                return (
-                  <tr key={idx}>
-                    <td>{idx + 1}</td>
-                    <td>
-                      <Markdown overrideStyle={false}>
-                        {`==(${combo.mods.join(' + ')}) x 100%[2f]==`}
-                      </Markdown>
-                    </td>
-                    <td>{combo.mods.length}</td>
-                    <td>{combo.sp}</td>
-                    {branchedChain.hasUtp && <td>{combo.utp}</td>}
-                    <td>
-                      <Markdown overrideStyle={false}>
-                        {`==(${combo.odRate.join(' + ')}) / ${combo.odRate.length}[2f]==`}
-                      </Markdown>
-                    </td>
-                    {
-                      branchedChain.hasCrisis &&
-                      <td>
-                        <Markdown overrideStyle={false}>
-                          {`==(${combo.crisisMod.join(' + ')}) / ${combo.crisisMod.length}[2f]==`}
-                        </Markdown>
-                      </td>
-                    }
-                    {
-                      combo.cancelToNextSec &&
-                      <>
-                        <td>{combo.cancelToNextSec.toFixed(2)}</td>
-                        <td>{(combo.sp / combo.cancelToNextSec).toFixed(2)}</td>
-                      </>
-                    }
-                  </tr>
-                );
-              })
-            }
-            {
-              <tr>
-                <td>{t((t) => t.game.unitInfo.text.total)}</td>
-                <td>
-                  {(sum(branchedChain.combos.map((combo) => sum(combo.mods))) * 100).toFixed(2)}&nbsp;%
-                </td>
-                <td>
-                  {sum(branchedChain.combos.map((combo) => combo.mods.length)).toFixed(0)}
-                </td>
-                <td>
-                  {sum(branchedChain.combos.map((combo) => combo.sp)).toFixed(0)}
-                </td>
-                {
-                  branchedChain.hasUtp &&
+            {branchedChain.combos.map((combo, idx) => {
+              return (
+                <tr key={idx}>
+                  <td>{idx + 1}</td>
                   <td>
-                    {sum(branchedChain.combos.map((combo) => combo.utp)).toFixed(2)}
+                    <Markdown overrideStyle={false}>
+                      {`==(${combo.mods.join(' + ')}) x 100%[2f]==`}
+                    </Markdown>
                   </td>
-                }
+                  <td>{combo.mods.length}</td>
+                  <td>{combo.sp}</td>
+                  {branchedChain.hasUtp && <td>{combo.utp}</td>}
+                  <td>
+                    <Markdown overrideStyle={false}>
+                      {`==(${combo.odRate.join(' + ')}) / ${combo.odRate.length}[2f]==`}
+                    </Markdown>
+                  </td>
+                  {
+                    branchedChain.hasCrisis &&
+                    <td>
+                      <Markdown overrideStyle={false}>
+                        {`==(${combo.crisisMod.join(' + ')}) / ${combo.crisisMod.length}[2f]==`}
+                      </Markdown>
+                    </td>
+                  }
+                  {
+                    combo.cancelToNextSec &&
+                    <>
+                      <td>{combo.cancelToNextSec.toFixed(2)}</td>
+                      <td>{(combo.sp / combo.cancelToNextSec).toFixed(2)}</td>
+                    </>
+                  }
+                </tr>
+              );
+            })}
+            <tr>
+              <td>{t((t) => t.game.unitInfo.text.total)}</td>
+              <td>
+                {(sum(branchedChain.combos.map((combo) => sum(combo.mods))) * 100).toFixed(2)}&nbsp;%
+              </td>
+              <td>
+                {sum(branchedChain.combos.map((combo) => combo.mods.length)).toFixed(0)}
+              </td>
+              <td>
+                {sum(branchedChain.combos.map((combo) => combo.sp)).toFixed(0)}
+              </td>
+              {
+                branchedChain.hasUtp &&
+                <td>
+                  {sum(branchedChain.combos.map((combo) => combo.utp)).toFixed(2)}
+                </td>
+              }
+              <td>-</td>
+              {
+                branchedChain.hasCrisis &&
                 <td>-</td>
-                {
-                  branchedChain.hasCrisis &&
-                  <td>-</td>
-                }
-                <td>
-                  {sum(branchedChain.combos.map((combo) => combo.cancelToNextSec || 0)).toFixed(2)}
-                </td>
-                <td>
-                  {(
-                    sum(branchedChain.combos.map((combo) => combo.sp)) /
-                    sum(branchedChain.combos.map((combo) => combo.cancelToNextSec || 0))
-                  ).toFixed(2)}
-                </td>
-              </tr>
-            }
+              }
+              <td>
+                {sum(branchedChain.combos.map((combo) => combo.cancelToNextSec || 0)).toFixed(2)}
+              </td>
+              <td>
+                {(
+                  sum(branchedChain.combos.map((combo) => combo.sp)) /
+                  sum(branchedChain.combos.map((combo) => combo.cancelToNextSec || 0))
+                ).toFixed(2)}
+              </td>
+            </tr>
           </tbody>
         </table>
       </div>
